generative_UI_project_example/second-brain-research-dashboard/agent/pyproject.toml
[project]
name = "second-brain-agent"
version = "0.1.0"
description = "Pydantic AI agent for Second Brain Dashboard - transforms Markdown research into generative UI dashboards"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic-ai>=0.0.1",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "ag-ui-protocol>=0.1.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0.0",
    "python-dateutil>=2.8.0",
    "copilotkit>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "black>=24.0.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
